Внесены правки
